requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
aiogram>=3.4.0
logging.handlers
# Тесты
pytest>=7.4.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
coverage>=7.0.0

# Опционально для CLI: prompt_toolkit>=3.0.0
